+++ test/assets/fixtures/_base/js/lib/base-lib1.js
@@ -1,0 +1,3 @@
+(function baseLib1() {
+	console.log(arguments.callee.name);
+})();