+++ test/terrific-modules/config-terrific.js
@@ -5,7 +5,7 @@ module.exports = {
 	// file system paths relative to app.js, a property called app.config.pathsAbsolute will be generated from them
 	,paths: {
 		// path to the Terrific modules directories
-		modulesBaseDir: 'test/fixtures/terrific-modules'
+		modulesBaseDir: 'test/terrific-modules/fixtures'
 	}
 	,moduleDirName:         'mod-{{name}}'
 };