+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Add option to allow bypassing auth for configurable IP ranges.
  - Renamed app_modules to lib.
  - Move routes.js from lib to controllers.
+ - Test grunt tasks: build-external-js, build-external-css, tests now contained in subdirs
 
 # 0.6.2